+++ crates/minerva/src/attribute_store.rs
@@ -74,6 +74,9 @@ impl Change for AddAttributes {
     async fn apply(&self, client: &mut Client) -> ChangeResult {
         let tx = client.transaction().await?;
 
+        tx.execute("SET citus.multi_shard_modify_mode TO 'sequential'", &[])
+            .await?;
+
         let query = concat!(
             "SELECT attribute_directory.create_attribute(attribute_store, $1::name, $2::text, $3::text) ",
             "FROM attribute_directory.attribute_store ",
@@ -351,6 +354,9 @@ impl fmt::Display for AddAttributeStore {
 impl Change for AddAttributeStore {
     async fn apply(&self, client: &mut Client) -> ChangeResult {
         let tx = client.transaction().await?;
+
+        tx.execute("SET citus.multi_shard_modify_mode TO 'sequential'", &[])
+            .await?;
 
         let query = concat!(
             "CALL attribute_directory.create_attribute_store(",