+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Library: Use Citus sequential mode for attribute store changes to prevent
+  Citus related errors. 
+
 ## [9.24.0] - 2025-03-10
 
 ### Changed